02_Sales Analytics Platform/requirements.txt
# Sales Dashboard Requirements
pandas==2.2.3
numpy==1.26.4
plotly==5.24.1
dash==2.18.2
dash-bootstrap-components==1.6.0
scikit-learn==1.5.2
pytest==8.3.4
pyarrow==18.1.0

# Jupyter Notebooks for Interactive Analysis
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Real-time Streaming & AI
python-socketio>=5.11.0
uvicorn>=0.30.0
fastapi>=0.110.0
openai>=1.12.0
